feat: Add call-to-action button to services.html
At the end of the services.html page we want to funnel the potential
customers to the contact form to become a potential lead.

This commit adds a Call of action button at the bottom of the page.
The orange button is also renamed in the CSS file to be DRY and reusable