fix: Change links for the homepage to be index.html
As no HTML base tag is currently used, using / to access the homepage does not
work within GitHub pages. This results in 404 errors when not in a
development enviroment.

This commit changes the homepage links to be index.html to ensure this
is compatible with GitHub pages and running within a folder.